feat(children): bones with nested layout: children :rocket:
Extracted Bones component, ready to become a recursive component that prints children.
